refactor: make getHyphenIdsFromCurrentPage be module
